getting exchangeRates with console command

+++ src/Entity/ExchangeRate.php
@@ -23,7 +23,7 @@ class ExchangeRate
     private $type;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $rate;
 
@@ -76,12 +76,12 @@ class ExchangeRate
         return $this;
     }
 
-    public function getRate(): ?int
+    public function getRate(): ?float
     {
         return $this->rate;
     }
 
-    public function setRate(int $rate): self
+    public function setRate(float $rate): self
     {
         $this->rate = $rate;
